Update ShoppingCartPage, create CartItem
