+++ src/main/resources/dev.properties
@@ -1,4 +1,5 @@
 Site=http\://184-dp.tk
+#Site = https://dp184taqc.000webhostapp.com/
 AdminSite=http\://184-dp.tk/admin
-Driver=webdriver.gecko.driver
-PathForDriver=C\:\\Users\\alexv\\Downloads\\Java\\drivers\\geckodriver.exe
+Driver=webdriver.chrome.driver
+PathForDriver=C:\\Users\\Anna\\WebDriver\\chromedriver.exe
